refactor: decompose monolithic app.component.ts into modular screens
Extracts route-level components (Dashboard, LiveRun, Chat, Builder, Tools, Memory, Providers, Policy) from apps/magnetar-ui/src/app/app.component.ts into dedicated subdirectories within apps/magnetar-ui/src/app/screens/. Each screen now has separate .ts and .html files (and .css where applicable) to separate view concerns from state and orchestration logic, addressing scalability issues and enabling independent testability. app.routes.ts was updated to reference the new component locations.
