+++ apps/magnetar-ui/src/app/app.component.css
@@ -1,14 +1,27 @@
-.container {
-  max-width: 720px;
-  margin: 2rem auto;
-  padding: 1rem;
-  font-family: Inter, Arial, sans-serif;
+.custom-scrollbar::-webkit-scrollbar {
+  width: 8px;
+  height: 8px;
 }
-
-h1 {
-  margin-bottom: 0.5rem;
+.custom-scrollbar::-webkit-scrollbar-track {
+  background: transparent;
 }
-
-p {
-  line-height: 1.5;
+.custom-scrollbar::-webkit-scrollbar-thumb {
+  background: rgba(255, 255, 255, 0.1);
+  border-radius: 4px;
+}
+.custom-scrollbar::-webkit-scrollbar-thumb:hover {
+  background: rgba(255, 255, 255, 0.2);
+}
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+    transform: translateY(10px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+.animate-fade-in {
+  animation: fadeIn 0.4s ease-out forwards;
 }
