+++ apps/magnetar-ui/src/app/app.component.html
@@ -1,0 +1,30 @@
+<div class="min-h-screen bg-[#050505] text-zinc-300 font-sans flex flex-col selection:bg-cyan-500/30 selection:text-cyan-100 relative overflow-hidden">
+  <header class="h-14 border-b border-white/5 bg-[#0a0a0d]/80 backdrop-blur-md flex items-center justify-between px-4 z-50 sticky top-0">
+    <div class="flex items-center gap-8">
+      <div class="flex items-center gap-3">
+        <div class="w-8 h-8 rounded bg-gradient-to-br from-cyan-400 to-violet-600 flex items-center justify-center shadow-[0_0_15px_rgba(56,189,248,0.3)]">
+          <div class="w-3 h-3 bg-[#050505] rounded-full"></div>
+        </div>
+        <span class="font-bold text-lg tracking-tight text-white shadow-sm">MAGNETAR<span class="font-light text-zinc-400 ml-1">EIDOLON</span></span>
+      </div>
+      <nav class="hidden md:flex gap-1 bg-white/[0.03] p-1 rounded-lg border border-white/5">
+        <ng-container *ngFor="let tab of tabs">
+          <a
+            [routerLink]="tab.route"
+            routerLinkActive="bg-white/10 text-white shadow-sm"
+            #activeLink="routerLinkActive"
+            [routerLinkActiveOptions]="{ exact: true }"
+            [ngClass]="activeLink.isActive ? 'bg-white/10 text-white shadow-sm' : 'text-zinc-400 hover:text-zinc-200 hover:bg-white/5'"
+            class="px-4 py-1.5 rounded-md text-sm font-medium transition-all flex items-center gap-2">
+            <ui-icon [name]="tab.icon" [size]="14"></ui-icon>
+            {{ tab.label }}
+          </a>
+        </ng-container>
+      </nav>
+    </div>
+  </header>
+
+  <main class="flex-1 overflow-auto p-6 relative z-10 custom-scrollbar">
+    <router-outlet></router-outlet>
+  </main>
+</div>
