+++ apps/magnetar-ui/src/app/screens/builder/builder-screen.component.html
@@ -1,0 +1,22 @@
+<div class="h-[calc(100vh-6rem)] grid grid-cols-4 gap-0 border border-white/5 rounded-xl overflow-hidden animate-fade-in shadow-2xl bg-[#08080a]">
+  <!-- ... (keeping existing template as is) ... -->
+  <div class="col-span-3 relative overflow-hidden bg-[radial-gradient(ellipse_at_center,_var(--tw-gradient-stops))] from-white/[0.03] to-transparent">
+    <div class="absolute inset-0 bg-[url('data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNDAiIGhlaWdodD0iNDAiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PGNpcmNsZSBjeD0iMSIgY3k9IjEiIHI9IjEiIGZpbGw9InJnYmEoMjU1LDI1NSwyNTUsMC4wNSkiLz48L3N2Zz4=')]"></div>
+    <div class="absolute top-4 left-4 bg-[#121216] border border-white/10 rounded-lg p-1.5 flex gap-1 z-20 shadow-lg backdrop-blur-md">
+      <button class="px-3 py-1.5 bg-white/10 rounded text-xs font-medium text-white shadow-sm">Visual</button>
+      <button class="px-3 py-1.5 hover:bg-white/5 rounded text-xs font-medium text-zinc-400 transition-colors">YAML</button>
+    </div>
+    <div class="absolute inset-0 z-10 flex items-center justify-center pointer-events-none"></div>
+  </div>
+  <div class="col-span-1 border-l border-white/5 bg-[#0a0a0d] flex flex-col h-full">
+    <div class="p-4 border-b border-white/5 flex items-center gap-3 bg-white/[0.02]">
+      <div class="w-8 h-8 rounded bg-white/10 flex items-center justify-center">
+        <ui-icon name="cpu" [size]="16" cssClass="text-white"></ui-icon>
+      </div>
+      <div>
+        <div class="text-sm font-medium text-zinc-200">System Prompt Node</div>
+        <div class="text-xs text-zinc-500">ID: node_a7x9</div>
+      </div>
+    </div>
+  </div>
+</div>
