+++ apps/magnetar-ui/src/app/screens/live-run/live-run-screen.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { UiBadgeComponent } from '../../ui/badge.component.js';
+import { UiIconComponent } from '../../ui/icon.component.js';
+
+@Component({
+  selector: 'screen-live-run',
+  standalone: true,
+  imports: [CommonModule, UiIconComponent, UiBadgeComponent],
+  templateUrl: './live-run-screen.component.html',
+})
+export class LiveRunScreen {}
