+++ apps/magnetar-ui/src/app/screens/providers/providers-screen.component.css
@@ -1,0 +1,11 @@
+.accordion-grid {
+    display: grid;
+    grid-template-rows: 0fr;
+    transition: grid-template-rows 0.3s ease-in-out;
+}
+.accordion-grid.is-open {
+    grid-template-rows: 1fr;
+}
+.accordion-inner {
+    overflow: hidden;
+}
