+++ apps/magnetar-ui/src/app/screens/tools/tools-screen.component.html
@@ -1,0 +1,23 @@
+<div class="space-y-6 animate-fade-in pb-12">
+  <div class="flex justify-between items-end">
+    <div>
+      <h2 class="text-2xl font-light text-white tracking-tight">Tool Catalog</h2>
+      <p class="text-sm text-zinc-400 mt-1">Discover, configure, and govern capabilities for your agents.</p>
+    </div>
+  </div>
+
+  <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4">
+    <div *ngFor="let tool of tools" class="bg-[#0f0f13] border border-white/5 rounded-xl p-5 hover:border-white/20 transition-all hover:shadow-xl group relative overflow-hidden flex flex-col h-48">
+      <div class="flex justify-between items-start mb-4 relative z-10">
+        <div class="w-10 h-10 rounded-lg bg-white/5 border border-white/10 flex items-center justify-center text-zinc-300">
+          <ui-icon [name]="tool.icon" [size]="20"></ui-icon>
+        </div>
+        <ui-badge [status]="getToolStatusBadge(tool)">{{ tool.status }}</ui-badge>
+      </div>
+      <div class="relative z-10 flex-1">
+        <h3 class="font-medium text-zinc-200 text-base group-hover:text-white transition-colors">{{ tool.name }}</h3>
+        <p class="text-xs text-zinc-500 mt-1">{{ tool.category }}</p>
+      </div>
+    </div>
+  </div>
+</div>
